% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{compare_faces}
\alias{compare_faces}
\title{Compares a face in the \emph{source} input image with each of the 100 largest faces detected in the \emph{target} input image}
\usage{
compare_faces(SourceImage, TargetImage, SimilarityThreshold = NULL)
}
\arguments{
\item{SourceImage}{The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.}

\item{TargetImage}{The target image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.}

\item{SimilarityThreshold}{The minimum level of confidence in the face matches that a match must meet to be included in the \code{FaceMatches} array.}
}
\description{
Compares a face in the \emph{source} input image with each of the 100 largest faces detected in the \emph{target} input image.
}
\details{
If the source image contains multiple faces, the service detects the largest face and compares it with each face detected in the target image.

You pass the input and target images either as base64-encoded image bytes or as references to images in an Amazon S3 bucket. If you use the AWS CLI to call Amazon Rekognition operations, passing image bytes isn't supported. The image must be formatted as a PNG or JPEG file.

In response, the operation returns an array of face matches ordered by similarity score in descending order. For each face match, the response provides a bounding box of the face, facial landmarks, pose details (pitch, role, and yaw), quality (brightness and sharpness), and confidence value (indicating the level of confidence that the bounding box contains a face). The response also provides a similarity score, which indicates how closely the faces match.

By default, only faces with a similarity score of greater than or equal to 80% are returned in the response. You can change this value by specifying the \code{SimilarityThreshold} parameter.

\code{CompareFaces} also returns an array of faces that don't match the source image. For each face, it returns a bounding box, confidence value, landmarks, pose details, and quality. The response also returns information about the face in the source image, including the bounding box of the face and confidence value.

If the image doesn't contain Exif metadata, \code{CompareFaces} returns orientation information for the source and target images. Use these values to display the images with the correct image orientation.

If no faces are detected in the source or target images, \code{CompareFaces} returns an \code{InvalidParameterException} error.

This is a stateless API operation. That is, data returned by this operation doesn't persist.

For an example, see Comparing Faces in Images in the Amazon Rekognition Developer Guide.

This operation requires permissions to perform the \code{rekognition:CompareFaces} action.
}
\examples{
# This operation compares the largest face detected in the source image
# with each face detected in the target image.
compare_faces(
  SimilarityThreshold = 90L,
  SourceImage = list(
    S3Object = list(
      Bucket = "mybucket",
      Name = "mysourceimage"
    )
  ),
  TargetImage = list(
    S3Object = list(
      Bucket = "mybucket",
      Name = "mytargetimage"
    )
  )
)

}
