% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{create_collection}
\alias{create_collection}
\title{Creates a collection in an AWS Region}
\usage{
create_collection(CollectionId)
}
\arguments{
\item{CollectionId}{ID for the collection that you are creating.}
}
\description{
Creates a collection in an AWS Region. You can add faces to the collection using the operation.
}
\details{
For example, you might create collections, one for each of your application users. A user can then index faces using the \code{IndexFaces} operation and persist results in a specific collection. Then, a user can search the collection for faces in the user-specific container.

Collection names are case-sensitive.

This operation requires permissions to perform the \code{rekognition:CreateCollection} action.
}
\examples{
# This operation creates a Rekognition collection for storing image data.
create_collection(
  CollectionId = "myphotos"
)

}
