% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{delete_collection}
\alias{delete_collection}
\title{Deletes the specified collection}
\usage{
delete_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] ID of the collection to delete.}
}
\description{
Deletes the specified collection. Note that this operation removes all faces in the collection. For an example, see delete-collection-procedure.
}
\details{
This operation requires permissions to perform the \code{rekognition:DeleteCollection} action.
}
\section{Accepted Parameters}{
\preformatted{delete_collection(
  CollectionId = "string"
)
}
}

\examples{
# This operation deletes a Rekognition collection.
\donttest{delete_collection(
  CollectionId = "myphotos"
)}

}
