% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{delete_stream_processor}
\alias{delete_stream_processor}
\title{Deletes the stream processor identified by Name}
\usage{
delete_stream_processor(Name)
}
\arguments{
\item{Name}{[required] The name of the stream processor you want to delete.}
}
\description{
Deletes the stream processor identified by \code{Name}. You assign the value for \code{Name} when you create the stream processor with CreateStreamProcessor. You might not be able to use the same name for a stream processor for a few seconds after calling \code{DeleteStreamProcessor}.
}
\section{Accepted Parameters}{
\preformatted{delete_stream_processor(
  Name = "string"
)
}
}

