% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{describe_collection}
\alias{describe_collection}
\title{Describes the specified collection}
\usage{
describe_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] The ID of the collection to describe.}
}
\description{
Describes the specified collection. You can use \code{DescribeCollection} to get information, such as the number of faces indexed into a collection and the version of the model used by the collection for face detection.
}
\details{
For more information, see Describing a Collection in the Amazon Rekognition Developer Guide.
}
\section{Accepted Parameters}{
\preformatted{describe_collection(
  CollectionId = "string"
)
}
}

