% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{describe_stream_processor}
\alias{describe_stream_processor}
\title{Provides information about a stream processor created by}
\usage{
describe_stream_processor(Name)
}
\arguments{
\item{Name}{Name of the stream processor for which you want information.}
}
\description{
Provides information about a stream processor created by . You can get information about the input and output streams, the input parameters for the face recognition being performed, and the current status of the stream processor.
}
\examples{

}
