% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{get_celebrity_info}
\alias{get_celebrity_info}
\title{Gets the name and additional information about a celebrity based on his or her Amazon Rekognition ID}
\usage{
get_celebrity_info(Id)
}
\arguments{
\item{Id}{The ID for the celebrity. You get the celebrity ID from a call to the operation, which recognizes celebrities in an image.}
}
\description{
Gets the name and additional information about a celebrity based on his or her Amazon Rekognition ID. The additional information is returned as an array of URLs. If there is no additional information about the celebrity, this list is empty.
}
\details{
For more information, see Recognizing Celebrities in an Image in the Amazon Rekognition Developer Guide.

This operation requires permissions to perform the \code{rekognition:GetCelebrityInfo} action.
}
\examples{

}
