% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{get_label_detection}
\alias{get_label_detection}
\title{Gets the label detection results of a Amazon Rekognition Video analysis started by}
\usage{
get_label_detection(JobId, MaxResults = NULL, NextToken = NULL,
  SortBy = NULL)
}
\arguments{
\item{JobId}{[required] Job identifier for the label detection operation for which you want results returned. You get the job identifer from an initial call to \code{StartlabelDetection}.}

\item{MaxResults}{Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.}

\item{NextToken}{If the previous response was incomplete (because there are more labels to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of labels.}

\item{SortBy}{Sort to use for elements in the \code{Labels} array. Use \code{TIMESTAMP} to sort array elements by the time labels are detected. Use \code{NAME} to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by \code{TIMESTAMP}.}
}
\description{
Gets the label detection results of a Amazon Rekognition Video analysis started by .
}
\details{
The label detection operation is started by a call to which returns a job identifier (\code{JobId}). When the label detection operation finishes, Amazon Rekognition publishes a completion status to the Amazon Simple Notification Service topic registered in the initial call to \code{StartlabelDetection}. To get the results of the label detection operation, first check that the status value published to the Amazon SNS topic is \code{SUCCEEDED}. If so, call and pass the job identifier (\code{JobId}) from the initial call to \code{StartLabelDetection}.

\code{GetLabelDetection} returns an array of detected labels (\code{Labels}) sorted by the time the labels were detected. You can also sort by the label name by specifying \code{NAME} for the \code{SortBy} input parameter.

The labels returned include the label name, the percentage confidence in the accuracy of the detected label, and the time the label was detected in the video.

Use MaxResults parameter to limit the number of labels returned. If there are more results than specified in \code{MaxResults}, the value of \code{NextToken} in the operation response contains a pagination token for getting the next set of results. To get the next page of results, call \code{GetlabelDetection} and populate the \code{NextToken} request parameter with the token value returned from the previous call to \code{GetLabelDetection}.

\code{GetLabelDetection} doesn't return a hierarchical taxonomy, or bounding box information, for detected labels. \code{GetLabelDetection} returns \code{null} for the \code{Parents} and \code{Instances} attributes of the object which is returned in the \code{Labels} array.
}
\section{Accepted Parameters}{
\preformatted{get_label_detection(
  JobId = "string",
  MaxResults = 123,
  NextToken = "string",
  SortBy = "NAME"|"TIMESTAMP"
)
}
}

