% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_package.R
\docType{package}
\name{paws.rekognition-package}
\alias{paws.rekognition}
\alias{paws.rekognition-package}
\title{paws.rekognition: Amazon Rekognition}
\description{
This is the Amazon Rekognition API reference.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
