% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rekognition_operations.R
\name{stop_stream_processor}
\alias{stop_stream_processor}
\title{Stops a running stream processor that was created by CreateStreamProcessor}
\usage{
stop_stream_processor(Name)
}
\arguments{
\item{Name}{[required] The name of a stream processor created by CreateStreamProcessor.}
}
\description{
Stops a running stream processor that was created by CreateStreamProcessor.
}
\section{Accepted Parameters}{
\preformatted{stop_stream_processor(
  Name = "string"
)
}
}

